flask-web-app/requirements.txt
Flask==0.12.2
numpy==1.14.2
pillow==5.0.0
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0